create a smoke test for user app flow
